# documentation/api-openapi.yml
openapi: 3.0.1
info:
  title: Productivity API
  version: v1
servers:
  - url: 'http://localhost:8018'
    description: Generated server url
paths:
...
  '/user/{userId}/tasks':
    post:
      tags:
        - task-controller
      operationId: createTask
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Task Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        titleTask:
          type: string
          example: "Create template Go app"
        descriptionTask:
          type: string
          example: "Want to create a Go app that I can reuse over and over to quickly get a project up and running"
        priorityTask:
          type: string
          enum:
            - "high"
            - "medium"
            - "low"
        estimatedTimeTask:
          type: integer
          format: int32
          example: 75
        typeTask:
          type: string
          enum:
            - "physical"
            - "spiritual"
            - "emotional"
            - "intellectual"
        statusTask:
          type: string
          enum:
            - "backlog"
            - "to-do"
            - "in-progress"
            - "blocked"
            - "done"
    TaskResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/Task'